fixed probably unwanted integer division

+++ src/core/crop-module.cpp
@@ -4511,8 +4511,8 @@ monica::cropmodule::calcRootDensityFactorAndSum(const CropModule *cm) {
     } else if (i_Layer < cm->vc_RootingZone) {
       vc_RootDensityFactor[i_Layer] =
           exp(-cm->pc_RootFormFactor * (i_Layer * layerThickness)) *
-          (1.0 - ((i_Layer - cm->vc_RootingDepth) /
-                  (cm->vc_RootingZone - cm->vc_RootingDepth))); // []
+          (1.0 - (double(i_Layer - cm->vc_RootingDepth) /
+                  double(cm->vc_RootingZone - cm->vc_RootingDepth))); // []
     } else {
       vc_RootDensityFactor[i_Layer] = 0.0;
     } // []
